+++ apps/cron/src/core/logger.ts
@@ -1,6 +1,16 @@
 export class Logger {
   private static instance: Logger;
   private messages: string[] = [];
+  private _hasWarning = false;
+  private _hasError = false;
+
+  public get hasWarning() {
+    return this._hasWarning;
+  }
+
+  public get hasError() {
+    return this._hasError;
+  }
 
   public static getInstance(): Logger {
     if (!Logger.instance) {
@@ -18,9 +28,16 @@ export class Logger {
     console.log(message);
   }
 
+  public warn(message: string) {
+    this.messages.push(message);
+    console.warn(message);
+    this._hasWarning = true;
+  }
+
   public error(message: string) {
     this.messages.push(message);
     console.error(message);
+    this._hasError = true;
   }
 
   public getMessages() {